Tilpassinger til design, slått sammen controllere etc.
git-svn-id: https://subversion.kantega.no/svn/aksess/projects/aksess-modules@2608 497bca9e-fae7-0310-9681-f2327195b1a6

+++ modules/forum/src/java/no/kantega/forum/permission/PermissionManager.java
@@ -1,5 +1,5 @@
 package no.kantega.forum.permission;
 
 public interface PermissionManager {
-    public boolean hasPermission(String user, int permission, Object object);
+    public boolean hasPermission(String user, long permission, Object object);
 }
